requirements.txt
opencv-python
mediapipe
numpy
pyserial
pyyaml
pytest
